Fixed bad type
git-svn-id: file:///Users/min/Documents/svnrepros/svn_mindia/trunk@578 5f0a215f-9dc6-43d0-aa08-c0b9dc37ed8b

+++ src/misctools.cpp
@@ -373,7 +373,7 @@ bool QImageCache::Add( const QString & sImageFileName )
             aImage = aImage.scaled(m_iMaxWidth,m_iMaxHeight,Qt::KeepAspectRatio);
         }
 
-        m_aMap[sImageFileName] = QPair<QImage,unsigned long>(aImage,1);
+        m_aMap[sImageFileName] = QPair<QImage,int>(aImage,1);
         PushToLastImageFileNames( sImageFileName );
         m_bModified = true;
     }
